change attr in class player.rb move prints/puts. Modify readme.
