+++ README.md
@@ -1,1 +1,23 @@
-# rubytic-tac-toe
+![](https://img.shields.io/badge/Microverse-blueviolet)
+# Ruby Tic-Tac-Toe
+> In this project we manage to create the classic game of tic-tac-toe in ruby.
+## Built With
+- Ruby
+- Rubocop
+- Git
+## How to run it
+To get a local copy up and running follow these simple steps.
+* If you're using latest version of git.
+* Clone this url https://github.com/Milypm/Ruby-Bubble-Sort/tree/main on your local host.
+* Un zip and open the folder that contains the file and run the command line
+ruby ./bin/main.rb 
+and follow the game flow.
+## Authors
+**Carlos Ospina**
+- GitHub: [@carloso0114](https://github.com/carloso0114)
+- LinkedIn: [Carlos Ospina](https://www.linkedin.com/in/carlos-ospina-242b831a6/)
+**Varun Sharma**
+- GitHub: [@Varun](https://github.com/Varun2323)
+- LinkedIn: [Varun](https://www.linkedin.com/in/varun-sharma-82b29b82/)
+## Show your support
+Give a star if you like this project!
