+++ lib/player.rb
@@ -1,5 +1,6 @@
 class Player
-  attr_accessor :name, :symbol
+  attr_accessor :name
+  attr_reader :symbol
 
   def initialize(name, symbol)
     @name = name
